@@ -1,25 +1,25 @@
 module github.com/ethereum/node-crawler
 
-go 1.20
+go 1.21
 
 require (
-	github.com/ethereum/go-ethereum v1.12.2
-	github.com/fjl/memsize v0.0.1
+	github.com/ethereum/go-ethereum v1.13.2
+	github.com/fjl/memsize v0.0.2
 	github.com/gorilla/mux v1.8.0
 	github.com/hashicorp/golang-lru v1.0.2
 	github.com/mattn/go-colorable v0.1.13
 	github.com/mattn/go-isatty v0.0.19
 	github.com/oschwald/geoip2-golang v1.9.0
-	github.com/pkg/errors v0.9.1
 	github.com/protolambda/zrnt v0.30.0
 	github.com/protolambda/ztyp v0.2.2
 	github.com/urfave/cli/v2 v2.25.7
-	modernc.org/sqlite v1.25.0
+	modernc.org/sqlite v1.26.0
 )
 
 require (
 	github.com/DataDog/zstd v1.5.2 // indirect
-	github.com/StackExchange/wmi v0.0.0-20180116203802-5d049714c4a6 // indirect
+	github.com/Microsoft/go-winio v0.6.1 // indirect
+	github.com/StackExchange/wmi v1.2.1 // indirect
 	github.com/VictoriaMetrics/fastcache v1.6.0 // indirect
 	github.com/beorn7/perks v1.0.1 // indirect
 	github.com/bits-and-blooms/bitset v1.7.0 // indirect
